Bugs

+++ LMS_CMS_Backend/LMS_CMS_BL/DTO/Accounting/EmployeeAccountingPut.cs
@@ -24,14 +24,14 @@ namespace LMS_CMS_BL.DTO.Accounting
         public long? Nationality { get; set; }
         public string? DateOfAppointment { get; set; }
         public string? DateOfLeavingWork { get; set; }
-        public int? MonthSalary { get; set; }
+        public decimal? MonthSalary { get; set; }
         public bool? HasAttendance { get; set; }
         public string? AttendanceTime { get; set; }
         public string? DepartureTime { get; set; }
         public float? DelayAllowance { get; set; }
-        public int? AnnualLeaveBalance { get; set; }
-        public int? CasualLeavesBalance { get; set; }
-        public int? MonthlyLeaveRequestBalance { get; set; }
+        public decimal? AnnualLeaveBalance { get; set; }
+        public decimal? CasualLeavesBalance { get; set; }
+        public decimal? MonthlyLeaveRequestBalance { get; set; }
         public int? GraduationYear { get; set; }
         public string? Note { get; set; }
         public long? AcademicDegreeID { get; set; }
